introduce registerServices
