+++ modules/nuxt-services/index.js
@@ -121,3 +121,21 @@ export default async function () {
     consola.info('Websockets server ready for services')
   })
 }
+
+export function registerServices (type, connectFunction) {
+  return async () => {
+    for (const service in this.options.services) {
+      const serviceData = this.options.services[service]
+      if (Array.isArray(serviceData) && serviceData[0] === 'postgresql') {
+        let settings = {}
+        if (!serviceData[1]) {
+          serviceData[1] = {}
+        }
+        if (Object.keys(serviceData).length === 0) {
+          throw new Error(`Configuration for \`${service}\` missing.`)
+        }
+        this.nuxt[`$${service}`] = await connect.apply(this, serviceData)
+      }
+    }
+  }
+}
